Handle gems installed with --user-install (#27)
When you install gems with `--user-install`, they will go in `~/.gem`
instead of in, e.g. `~/.rbenv` (or wherever your default GEM_PATH is).

This is a relatively uncommon way for gems to be installed, but it's
technically supported by bundler & rubygems, and it causes issues with
`uncruft`'s ability to resolve `$GEM_HOME` in ignorefile entries. This
PR adds an additional check against `Gem.user_dir` and treats it as
equivalent to `$GEM_HOME` (since the ignorefile is designed to work
across multiple systems, and should not care where the gems are
installed).